+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to Rust's notion of
 to 0.3.0 are beta releases.
 
 ## [Unreleased]
+### Changed
+- The "sharing violation" logic now also sends SIGHUP to any `yubikey-agent`
+  that is running, to have them release any YubiKey locks they are holding.
 
 ## [0.3.1] - 2022-12-30
 ### Changed